[refactor](be) Extract new parquet file context
### What problem does this PR solve?

Issue Number: None

Related PR: None

Problem Summary: Move the new parquet reader's Arrow random access file adapter and file-level Arrow Parquet reader lifecycle into ParquetFileContext, keeping ParquetReader focused on file-local scan orchestration.

### Release note

None

### Check List (For Author)

- Test: Manual test
    - Ran build-support/clang-format.sh for changed BE files
    - Ran git diff --check for changed BE files
- Behavior changed: No
- Does this need documentation: No
